returned JSON now has the capability of showing the original data types for properties.
element properties previously displayed as a name-value pair will now be option to contain a JSONObject with two properties "type" and "value".  the type will be one of: map, list, string, double, float, integer, long.

maps and lists work in a similar fashion where the list will be an array of JSONObjects and maps will be key value pairs with the value being a JSONObject.  in these cases, the JSONObjects will have the "type" and "value" combination.

added an new unit tests to cover this new serialization scheme.  fixed bugs in serialization of lists and maps that were probably introduced when jettison replaced simple-json.