Initial set up for the web tool that sits on top of Rexster.
Added javascript and html/css files.  Modified the build for maven to take these new resources into account.  The WebServer class now loads two instances of the Grizzly server.  Not sure if that is improper or not at the moment but it keeps the REST services from having to shift to a sub-directory.  Rexster still runs on 8182 and the new webserver that hosts the html content is on 8183.  Added some base functionality to add headers to the responses from the REST services.  Implemented it partially at this point, but probably needs to be extended to all service methods.

Corrected a bug in the parameters for configuration that was not allowing some of the basic configuration optiosn to work.