Added a mechanism to configure the url of the REST API that the dog house connects to.
Involved adding a base-uri config option to rexster.xml. This value is translated to a value that gets passed into the ToolServlet on an init parameter in the servlet context and is then loaded as a global variable in the javascript.

Also renamed a few of the configuration parameters in rexster.xml.  server-port = rexster-server-port and admin-server-port = doghouse-server-port.