Frames extension in play.
The basics of a frames extension.  Simply converts getter methods with a Property annotation and writes it out to JSON.  Didn't really look to build an all out frames extension...just wanted to use it to expand the extension configuration system which now has a different look in rexster.xml.  Configuration info is passed directly into the extension.