Removed unit tests that have been "ignored" for various reasons.
The bulk of the ignored tests were marked as "ignored" because of the adverse affect they were having on the build server.  They kept making it hang.  Since the last of these were marked as such there have been no reports of the build server tanking.  That was a few weeks ago.