The <graph-file> element in TinkerGraph configuration is now used locate a graph. #122
Prior to this change, the <graph-file> pointed to a GraphML file and the graph was loaded with the contents of that file.  It now points to the location of the graph.  The graph must already be initialized with data (ie. from a GraphML file).  The configuration no longer recognizes GraphML.