Made return of index meta data more consistent. #123 and #124
When requesting an index by name http://localhost/graph/indices/edge the index is wrapped in a results key in the return JSON.  This is more consistent with other return values.

Indices returned from http://localhost/graph/indices or when requested by name have their class value set to edge or vertex as opposed to the full canonical class name.