Supplied a Jersey Injectable Provider so that Rexster can be embedded.
Really didn't invent anything new here.  Original work was done by Jordan Lewis to get this injectable provider in play.  I removed that work in the grand refactoring on the way through 2.2.0 as it became less relevant for the standalone version of Rexster.  This puts the class back in so that it fits the current model more cleanly.