Refactor unit tests and remove need for self-xml reference.
the self-xml reference was injected into the rexster.xml at runtime and referred to the location of itself so that servlets could get that information as a configuration to configure graphs.  Since the graph list is now injected into the servlets programmatically there is no longer a need to have servlets self-configure.