Enabled application/x-www-form-urlencoded on extension routes #264
Can't remember the specifics of why this was taken out in the first place, but it's buried in issues or commit comments somewhere.  Had something to do with the parse to the JSON request object.  That whole thing needs to be simplified on its way to a next major version.  In any case, it probably makes sense to accept this mime type on extensions as it blocks folks from extending rexster in the way they want and that's basically the point of extensions in the first place.