Reset the script engine on rexster.xml and fix httpserver reset issues.
HTTP would undeploy servlets from all context on rexster.xml change.  Had to globally undeploy and globally redeploy to make it work.  A bit suboptimal in that there really is no need to undeploy/redeploy dog house if you just made a change to a REST setting.  Not sure if this is a bug in grizzly but it is how the code is written atm.