Adding hint-related logging statements
I've enclosed debug() and trace() calls on the potentially
performance-sensitive path in isDebugEnabled() and isTraceEnabled().
Some of the log strings are potentially costly to generate, and the
Log4j API that Rexster uses requires generating them even if trace and
debug are disabled (unless they are gated by these boolean checks).